refactor/FSTB-106: rename
